increase t-vals dataset precision

+++ analyses/searchlight/searchlight_permutation_testing.py
@@ -706,7 +706,7 @@ def calc_t_values_null_distr(args, out_path):
     with h5py.File(out_path, 'w') as all_t_vals_file:
         for hemi_metric in tmp_files[0].keys():
             tvals_shape = (args.n_permutations_group_level, n_vertices)
-            all_t_vals_file.create_dataset(hemi_metric, tvals_shape, dtype='float16', fillvalue=np.nan)
+            all_t_vals_file.create_dataset(hemi_metric, tvals_shape, dtype='float32', fillvalue=np.nan)
 
         for i in tqdm(range(args.n_permutations_group_level)):
             for hemi_metric in tmp_files[0].keys():
